support using PyHKEY as a context manager
This commit adds __enter__() and __exit__() functions to the PyHKEY class to
allow it to be used as a context manager, mimicking the standard winreg
library. The only thing this does is close the key at the end of the context
manager.

Testing is handled by modifying the key_handle_test_read_only() pytest fixture
to use the context manager.